requirements.txt
asgiref==3.7.2
Django==3.2.22
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
importlib-metadata==6.7.0
Markdown==3.4.4
Pillow==9.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
uuid==1.30
zipp==3.15.0
